refactor: routing structure

+++ src/pages/api/stream/index.ts
@@ -1,7 +1,9 @@
-import { client } from '@/lib/webtorrent'
+import client from '@/lib/webtorrent'
 import { Torrent } from 'webtorrent'
 import { NextApiRequest, NextApiResponse } from 'next'
 import fs from 'fs'
+import prisma from '@/lib/prisma'
+import logger from '@/lib/logger'
 
 export default async function handler(
   req: NextApiRequest,
@@ -20,20 +22,53 @@ export default async function handler(
       break;
     case 'DELETE':
       try {
-        for (const torrent of client.torrents) {
-          client.remove(torrent)
-          torrent.destroy({}, () => {
-            torrent.files.forEach((file) => {
-              console.log(file.path)
-              fs.unlinkSync(file.path)
-            })
-            fs.rmdirSync(torrent.path)
+        // for (const torrent of client.torrents) {
+        //   client.remove(torrent)
+        //   torrent.destroy({}, () => {
+        //     torrent.files.forEach((file) => {
+        //       console.log(file.path)
+        //       fs.unlinkSync(file.path)
+        //     })
+        //     fs.rmdirSync(torrent.path)
+        //   })
+        // }
+        const episodeNotConfirmed = await prisma.episode.findMany({
+          where: {
+            confirmed: false,
+            previewed: true,
+          },
+          select: {
+            id: true,
+            torrent: true,
+            infoHash: true,
+          }
+        })
+        let count = 0
+        if (episodeNotConfirmed) {
+          episodeNotConfirmed.forEach((episode) => {
+            if (episode.infoHash) {
+              count++
+            }
+            const isExist = client.get(episode.torrent)
+            const isExistTorrent = client
+            if (isExist) {
+              client.remove(isExist)
+              // delete torrent file
+              isExist.destroy({}, () => {
+                isExist.files.forEach((file) => {
+                  logger.info(file.path)
+                  fs.unlinkSync(file.path)
+                })
+                fs.rmdirSync(isExist.path)
+              })
+              count--
+            }
           })
         }
-
-        res.status(200).json({ success: true })
+        if (count !== 0) res.status(404).json({ message: `Cache not fully deleted` })
+        res.status(200).json({ message: `${count} torrents deleted` })
       } catch (e) {
-        res.status(500).json({ error: e.message })
+        res.status(500).json({ message: e.message })
       }
       break;
     default:
